// src/enums/ColumnsDetails.ts
export enum ColumnsDetails {
    Contact = 'contact',
    // Limit = 'limit',
    // Withdrawn = 'withdrawn',
    ChannelId = 'channelId',
    CurrentStatus = 'currentStatus',
    LoanText = 'loanText',
    SecondaryContact = 'secondaryContact',
    // ContactDesc = 'contactDesc',
    // SecondaryDesc = 'secondaryDesc',
    ChannelStatus = 'ChannelStatus',
    ContactData = 'contactData',
    SecondaryData = 'secondaryData',
    Company = 'company',
    Version = 'version'
}